fixed after review

+++ tests/moveFiles.test.js
@@ -12,25 +12,26 @@ const execAsync = util.promisify(exec);
 
 describe('File Move Tests', () => {
   const basePath = 'node src/app.js';
-  let testContent = '';
+  const testContent = faker.lorem.paragraphs();
   const testFileName = faker.system.commonFileName('txt');
 
   const tempDir = path.join('tests', faker.word.noun());
-  const testFile = path.join(tempDir, testFileName);
+  const testFilePath = path.join(tempDir, testFileName);
   const testDir = path.join('tests', faker.word.noun());
-
-  beforeAll(() => {
-    testContent = faker.lorem.paragraphs();
-  });
 
   beforeEach(() => {
     fs.mkdirSync(tempDir);
-    fs.writeFileSync(testFile, testContent);
+    fs.writeFileSync(testFilePath, testContent);
   });
 
   afterEach(() => {
-    fs.rmdirSync(tempDir, { recursive: true });
-    fs.rmdirSync(testDir, { recursive: true });
+    if (fs.existsSync(tempDir)) {
+      fs.rmdirSync(tempDir, { recursive: true });
+    }
+
+    if (fs.existsSync(testDir)) {
+      fs.rmdirSync(testDir, { recursive: true });
+    }
   });
 
   describe('without params', () => {
@@ -43,7 +44,7 @@ describe('File Move Tests', () => {
 
   describe('with one param', () => {
     test('should throw error', async() => {
-      const { stderr } = await execAsync(`${basePath} ${testFile}`);
+      const { stderr } = await execAsync(`${basePath} ${testFilePath}`);
 
       expect(stderr.length).toBeGreaterThan(0);
     });
@@ -53,7 +54,7 @@ describe('File Move Tests', () => {
     test('if source file does not exist, should throw error', async() => {
       const nonExistingFile = path.join(tempDir, faker.system.commonFileName('txt'));
 
-      const { stderr } = await execAsync(`${basePath} ${nonExistingFile} ${testFile}`);
+      const { stderr } = await execAsync(`${basePath} ${nonExistingFile} ${testFilePath}`);
 
       expect(stderr.length).toBeGreaterThan(0);
     });
@@ -61,50 +62,48 @@ describe('File Move Tests', () => {
     test('should rename a file, if destination is a new filename', async() => {
       const newFilePath = path.join(tempDir, faker.lorem.word());
 
-      const { stderr } = await execAsync(`${basePath} ${testFile} ${newFilePath}`);
+      const { stderr } = await execAsync(`${basePath} ${testFilePath} ${newFilePath}`);
 
       expect(stderr).toBeFalsy();
 
       const content = fs.readFileSync(newFilePath, 'utf-8');
 
       expect(fs.existsSync(newFilePath)).toBe(true);
-      expect(fs.existsSync(testFile)).toBe(false);
+      expect(fs.existsSync(testFilePath)).toBe(false);
       expect(content).toBe(testContent);
     });
 
     test('should do nothing if source and destination are the same', async() => {
-      const { stderr } = await execAsync(`${basePath} ${testFile} ${testFile}`);
+      const { stderr } = await execAsync(`${basePath} ${testFilePath} ${testFilePath}`);
 
-      const content = fs.readFileSync(testFile, 'utf-8');
+      const content = fs.readFileSync(testFilePath, 'utf-8');
 
       expect(stderr).toBeFalsy();
-      expect(fs.existsSync(testFile)).toBe(true);
+      expect(fs.existsSync(testFilePath)).toBe(true);
       expect(content).toBe(testContent);
     });
 
     test('should move file, if passed destination is a file without extension', async() => {
       const newFilePath = path.join(tempDir, faker.lorem.word());
-
-      const { stderr } = await execAsync(`${basePath} ${testFile} ${newFilePath}`);
+      const { stderr } = await execAsync(`${basePath} ${testFilePath} ${newFilePath}`);
 
       expect(stderr).toBeFalsy();
       expect(fs.existsSync(newFilePath)).toBe(true);
-      expect(fs.existsSync(testFile)).toBe(false);
+      expect(fs.existsSync(testFilePath)).toBe(false);
     });
 
     test('should move file, if passed destination is a directory', async() => {
       fs.mkdirSync(testDir);
 
-      const { stderr } = await execAsync(`${basePath} ${testFile} ${testDir}`);
+      const { stderr } = await execAsync(`${basePath} ${testFilePath} ${testDir}`);
 
       expect(stderr).toBeFalsy();
 
       const newPath = path.join(testDir, testFileName);
-
       const content = fs.readFileSync(newPath, 'utf-8');
 
       expect(fs.existsSync(newPath)).toBe(true);
-      expect(fs.existsSync(testFile)).toBe(false);
+      expect(fs.existsSync(testFilePath)).toBe(false);
       expect(content).toBe(testContent);
     });
 
@@ -112,24 +111,24 @@ describe('File Move Tests', () => {
       const nonExistingDir = path.join(tempDir, 'nonExistingDir', faker.word.noun());
 
       const { stderr } = await execAsync(
-        `${basePath} ${testFile} ${nonExistingDir}`
+        `${basePath} ${testFilePath} ${nonExistingDir}`
       );
 
       expect(stderr.length).toBeGreaterThan(0);
       expect(fs.existsSync(nonExistingDir)).toBe(false);
-      expect(fs.existsSync(testFile)).toBe(true);
+      expect(fs.existsSync(testFilePath)).toBe(true);
     });
 
     test('should throw error if destination is non-existed directory with fileName', async() => {
       const nonExistingDir = path.join(tempDir, 'nonExistingDir', faker.word.noun());
 
       const { stderr } = await execAsync(
-        `${basePath} ${testFile} ${nonExistingDir}`
+        `${basePath} ${testFilePath} ${nonExistingDir}`
       );
 
       expect(stderr.length).toBeGreaterThan(0);
       expect(fs.existsSync(nonExistingDir)).toBe(false);
-      expect(fs.existsSync(testFile)).toBe(true);
+      expect(fs.existsSync(testFilePath)).toBe(true);
     });
 
     test('should move file to directory path ending with "/" with the same filename', async() => {
@@ -137,17 +136,9 @@ describe('File Move Tests', () => {
 
       const newPath = path.join(testDir, '/');
 
-      await execAsync(`${basePath} ${testFile} ${newPath}`);
+      await execAsync(`${basePath} ${testFilePath} ${newPath}`);
 
       expect(fs.existsSync(path.join(newPath, testFileName))).toBe(true);
-    });
-
-    test('should move file into existing directory if destination is a directory', async() => {
-      fs.mkdirSync(testDir);
-
-      await execAsync(`${basePath} ${testFile} ${testDir}`);
-
-      expect(fs.existsSync(path.join(testDir, testFileName))).toBe(true);
     });
   });
 });
